Convert to a plugin
