+++ lib/index.js
@@ -1,36 +1,29 @@
+'use strict';
+
 // Load modules
 
-var Hapi = require('hapi');
-var Hoek = require('hoek');
-var Package = require('../package.json');
+const Hapi = require('hapi');
+const Hoek = require('hoek');
+const Version = require('./version');
 
 
 // Declare internals
 
-var internals = {};
+const internals = {};
 
 
 internals.init = function () {
 
-    var server = new Hapi.Server();
+    const server = new Hapi.Server();
     server.connection({ port: 8000 });
-
-    server.route({
-        method: 'GET',
-        path: '/version',
-        config: {
-            description: 'Returns the version of the server',
-            handler: function (request, reply) {
-
-                return reply({ version: Package.version });
-            }
-        }
-    });
-
-    server.start(function (err) {
+    server.register(Version, function (err) {
 
         Hoek.assert(!err, err);
-        console.log('Server started at: ' + server.info.uri);
+        server.start(function (err) {
+
+            Hoek.assert(!err, err);
+            console.log('Server started at: ' + server.info.uri);
+        });
     });
 };
